requirements.txt
celery[redis]
python-decouple
pymongo
